Backported changes to multiprocessing in Python trunk (Revision 78790)

+++ billiard/_pool.py
@@ -58,7 +58,9 @@ class SoftTimeLimitExceeded(Exception):
 def soft_timeout_sighandler(signum, frame):
     raise SoftTimeLimitExceeded()
 
-def worker(inqueue, outqueue, ackqueue, initializer=None, initargs=()):
+def worker(inqueue, outqueue, ackqueue, initializer=None, initargs=(),
+        maxtasks=None):
+    assert maxtasks is None or (type(maxtasks) == int and maxtasks > 0)
     pid = os.getpid()
     put = outqueue.put
     get = inqueue.get
@@ -73,7 +75,8 @@ def worker(inqueue, outqueue, ackqueue, initializer=None, initargs=()):
     if SIG_SOFT_TIMEOUT is not None:
         signal.signal(SIG_SOFT_TIMEOUT, soft_timeout_sighandler)
 
-    while 1:
+    completed = 0
+    while maxtasks is None or (maxtasks and completed < maxtasks):
         try:
             task = get()
         except (EOFError, IOError):
@@ -91,6 +94,9 @@ def worker(inqueue, outqueue, ackqueue, initializer=None, initargs=()):
         except Exception, e:
             result = (False, e)
         put((job, i, result))
+        completed += 1
+    debug('worker exiting after %d tasks' % completed)
+
 
 #
 # Class representing a process pool
@@ -103,13 +109,14 @@ class Pool(object):
     Process = Process
 
     def __init__(self, processes=None, initializer=None, initargs=(),
-            timeout=None, soft_timeout=None):
+            maxtasksperchild=None, timeout=None, soft_timeout=None):
         self._setup_queues()
         self._taskqueue = Queue.Queue()
         self._cache = {}
         self._state = RUN
         self.timeout = timeout
         self.soft_timeout = soft_timeout
+        self._maxtasksperchild = maxtasksperchild
         self._initializer = initializer
         self._initargs = initargs
 
@@ -122,7 +129,7 @@ class Pool(object):
                 processes = cpu_count()
             except NotImplementedError:
                 processes = 1
-        self._size = processes
+        self._processes = processes
 
         if initializer is not None and not hasattr(initializer, '__call__'):
             raise TypeError('initializer must be a callable')
@@ -130,6 +137,14 @@ class Pool(object):
         self._pool = []
         for i in range(processes):
             self._create_worker_process()
+
+        self._worker_handler = threading.Thread(
+            target=Pool._handle_workers,
+            args=(self, ),
+        )
+        self._worker_handler.daemon = True
+        self._worker_handler._state = RUN
+        self._worker_handler.start()
 
         self._task_handler = threading.Thread(
             target=Pool._handle_tasks,
@@ -176,7 +191,8 @@ class Pool(object):
             self, self._terminate_pool,
             args=(self._taskqueue, self._inqueue, self._outqueue,
                   self._ackqueue, self._pool, self._ack_handler,
-                  self._task_handler, self._result_handler, self._cache,
+                  self._worker_handler, self._task_handler,
+                  self._result_handler, self._cache,
                   self._timeout_handler, self._timeout_handler_stopped),
             exitpriority=15
             )
@@ -185,13 +201,42 @@ class Pool(object):
         w = self.Process(
             target=worker,
             args=(self._inqueue, self._outqueue, self._ackqueue,
-                    self._initializer, self._initargs)
+                    self._initializer, self._initargs, self._maxtasksperchild)
             )
         self._pool.append(w)
         w.name = w.name.replace('Process', 'PoolWorker')
         w.daemon = True
         w.start()
         return w
+
+    def _join_exited_workers(self):
+        """Cleanup after any worker processes which have exited due to
+        reaching their specified lifetime. Returns True if any workers were
+        cleaned up.
+        """
+        for i in reversed(range(len(self._pool))):
+            worker = self._pool[i]
+            if worker.exitcode is not None:
+                # worker exited
+                debug('cleaning up worker %d' % i)
+                worker.join()
+                del self._pool[i]
+        return len(self._pool) < self._processes
+
+    def _repopulate_pool(self):
+        """Bring the number of pool processes up to the specified number,
+        for use after reaping workers which have exited.
+        """
+        debug('repopulating pool')
+        for i in range(self._processes - len(self._pool)):
+            self._create_worker_process()
+            debug('added worker')
+
+    def _maintain_pool(self):
+        """"Clean up any exited workers and start replacements for them.
+        """
+        if self._join_exited_workers():
+            self._repopulate_pool()
 
     def _setup_queues(self):
         from multiprocessing.queues import SimpleQueue
@@ -297,6 +342,13 @@ class Pool(object):
         return result
 
     @staticmethod
+    def _handle_workers(pool):
+        while pool._worker_handler._state == RUN and pool._state == RUN:
+            pool._maintain_pool()
+            time.sleep(0.1)
+        debug('worker handler exiting')
+
+    @staticmethod
     def _handle_timeouts(pool, sentinel_event, cache, t_soft, t_hard):
         thread = threading.current_thread()
         processes = pool._pool
@@ -353,9 +405,8 @@ class Pool(object):
                 job._timeout_callback(soft=False)
             if not process:
                 return
-            # Terminate the process and create a new one.
+            # Terminate the process
             process.terminate()
-            pool._create_worker_process()
 
         # Inner-loop
         while 1:
@@ -377,6 +428,7 @@ class Pool(object):
             time.sleep(1) # Don't waste CPU cycles.
 
         debug('timeout handler exiting')
+
 
     @staticmethod
     def _handle_tasks(taskqueue, put, outqueue, pool):
@@ -543,20 +595,27 @@ class Pool(object):
         debug('closing pool')
         if self._state == RUN:
             self._state = CLOSE
+            self._worker_handler._state = CLOSE
             self._taskqueue.put(None)
 
     def terminate(self):
         debug('terminating pool')
         self._state = TERMINATE
+        self._worker_handler._state = TERMINATE
         self._terminate()
 
     def join(self):
-        debug('joining pool')
         assert self._state in (CLOSE, TERMINATE)
+        debug('!joining worker handler')
+        self._worker_handler.join()
+        debug('!joining task handler')
         self._task_handler.join()
+        debug('!joining result handler')
         self._result_handler.join()
+        debug('joining pool')
         for p in self._pool:
             p.join()
+        debug('after join()')
 
     @staticmethod
     def _help_stuff_finish(inqueue, task_handler, size):
@@ -569,10 +628,13 @@ class Pool(object):
 
     @classmethod
     def _terminate_pool(cls, taskqueue, inqueue, outqueue, ackqueue, pool,
-                        ack_handler, task_handler, result_handler, cache,
-                        timeout_handler, timeout_handler_stopped):
+                        ack_handler, worker_handler, task_handler,
+                        result_handler, cache, timeout_handler,
+                        timeout_handler_stopped):
         # this is guaranteed to only be called once
         debug('finalizing pool')
+
+        worker_handler._state = TERMINATE
 
         task_handler._state = TERMINATE
         taskqueue.put(None)                 # sentinel
@@ -592,10 +654,12 @@ class Pool(object):
             timeout_handler._state = TERMINATE
             timeout_handler_stopped.set()
 
+        # Terminate workers which haven't already finished
         if pool and hasattr(pool[0], 'terminate'):
             debug('terminating workers')
             for p in pool:
-                p.terminate()
+                if p.exitcode is None:
+                    p.terminate()
 
         debug('joining task handler')
         task_handler.join(1e100)
@@ -613,7 +677,10 @@ class Pool(object):
         if pool and hasattr(pool[0], 'terminate'):
             debug('joining pool workers')
             for p in pool:
-                p.join()
+                if p.is_alive():
+                    # worker has not yet exited
+                    debug('cleaning up worker %d' % p.pid)
+                    p.join()
 
 #
 # Class whose instances are returned by `Pool.apply_async()`
